ci(inttest): shutdown on crash
